Warn when we read more data than the file's declared size
Last commit changed how data is read for solid files, but this had a
side effect in that the data returned by the unrar lib was no longer
limited to the size of the file, as declared in its header.

Still keep this new behavior, but add a warning when this happens.