Basic GH CI for Windows
While there is already an AppVeyor CI, the problem with that is that
usually only one job runs in parallel, but we want to be able to run
the builds for multiple PHP versions.

For now, we just run for PHP 7.4, though, since some stuff needs to be
fixed, e.g. #13, and the failing tests[1].

[1] <https://github.com/cmb69/php-rar/runs/4370012547>